+++ R/compare_props.R
@@ -45,9 +45,10 @@ compare_props <- function(
   }
 
   ## check if there is variation in the data
-  if (any(summarise_all(dataset, does_vary) == FALSE)) {
-    return("One or more selected variables show no variation. Please select other variables." %>%
-      add_class("compare_props"))
+  not_vary <- colnames(dataset)[summarise_all(dataset, does_vary) == FALSE]
+  if (length(not_vary) > 0) {
+    return(paste0("The following variable(s) show no variation. Please select other variables.\n\n** ", paste0(not_vary, collapse = ", "), " **") %>%
+             add_class("compare_props"))
   }
 
   rn <- ""